Move SearchTokenizer from plume-models to plume-models::search::tokenizer
